requirements.txt
django~=3.0.3
mysqlclient~=2.0.1
djangorestframework~=3.11.1
ipython~=7.16.1